・材料画面の使用量を小数点で入力可能にした。 　・原材料の原価計算修正 　・コスト編集画面のタブをボタン型に変更 　・円、グラムを小数点第１位まで表示するよう修正 　・メイン画面、コスト編集画面でキャンセル操作した際に、編集有無をチェックして確認画面表示

+++ CostManager/Excel/CostReader.cs
@@ -37,11 +37,11 @@ namespace ExcelReaderUtility
             /// <summary>
             /// 原価
             /// </summary>
-            public uint cost;
+            public float cost;
             /// <summary>
             /// 原価基準容量
             /// </summary>
-            public uint gram;
+            public float gram;
         }
         public class WorkerData
         {
